Moved user models into seperate files
